+++ src/main/java/br/com/zalf/prolog/webservice/integracao/protheusrodalog/SistemaProtheusRodalog.java
@@ -6,6 +6,7 @@ import br.com.zalf.prolog.webservice.frota.pneu.afericao.model.NovaAfericaoPlaca
 import br.com.zalf.prolog.webservice.frota.veiculo.model.diagrama.DiagramaVeiculo;
 import br.com.zalf.prolog.webservice.integracao.IntegradorProLog;
 import br.com.zalf.prolog.webservice.integracao.protheusrodalog.model.NovaAfericaoPlacaProtheusRodalog;
+import br.com.zalf.prolog.webservice.integracao.protheusrodalog.model.ProtheusRodalogResponseAfericao;
 import br.com.zalf.prolog.webservice.integracao.protheusrodalog.model.TipoMedicaoAfericaoProtheusRodalog;
 import br.com.zalf.prolog.webservice.integracao.sistema.Sistema;
 import br.com.zalf.prolog.webservice.integracao.sistema.SistemaKey;
@@ -33,14 +34,23 @@ public class SistemaProtheusRodalog extends Sistema {
     @Override
     public Long insertAfericao(@NotNull final Long codUnidade, @NotNull final Afericao afericao) throws Throwable {
         final String tokenIntegracao = getIntegradorProLog().getTokenIntegracaoByCodUnidadeProLog(codUnidade);
-        return requester.insertAfericao(tokenIntegracao, codUnidade, ProtheusRodalogConverter.convertAfericao(afericao));
+        final ProtheusRodalogResponseAfericao responseAfericao = requester.insertAfericao(
+                tokenIntegracao,
+                codUnidade,
+                ProtheusRodalogConverter.convertAfericao(afericao));
+        if (responseAfericao.isStatus() && responseAfericao.getCodigoAfericaoInserida() != null) {
+            return responseAfericao.getCodigoAfericaoInserida();
+        } else {
+            throw new Exception("[INTEGRACAO - RODALOG] Não foi possível inserir a Aferição no sistema");
+        }
     }
 
     @NotNull
     @Override
     public CronogramaAfericao getCronogramaAfericao(@NotNull final Long codUnidade) throws Throwable {
         final String tokenIntegracao = getIntegradorProLog().getTokenIntegracaoByCodUnidadeProLog(codUnidade);
-        return ProtheusRodalogConverter.convertCronogramaAfericao(requester.getCronogramaAfericao(tokenIntegracao, codUnidade));
+        return ProtheusRodalogConverter
+                .convertCronogramaAfericao(requester.getCronogramaAfericao(tokenIntegracao, codUnidade));
     }
 
     @NotNull
@@ -56,13 +66,16 @@ public class SistemaProtheusRodalog extends Sistema {
                         placaVeiculo,
                         TipoMedicaoAfericaoProtheusRodalog.fromString(tipoAfericao));
         if (novaAfericaoPlaca.getCodDiagrama() == null) {
-            throw new IllegalStateException("Código do diagrama é null");
+            throw new IllegalStateException("[INTEGRACAO - RODALOG] O código do diagrama é null\n" +
+                    "CodUnidade: " + codUnidade + "\n" +
+                    "Placa: " + placaVeiculo);
         }
         final Optional<DiagramaVeiculo> diagramaVeiculo =
                 getIntegradorProLog().getDiagramaVeiculoByCodDiagrama(novaAfericaoPlaca.getCodDiagrama().shortValue());
         if (!diagramaVeiculo.isPresent()) {
             throw new IllegalStateException(
-                    "Nenhum diagrama encontrado para o código: " + novaAfericaoPlaca.getCodDiagrama());
+                    "[INTEGRACAO - RODALOG] Nenhum diagrama encontrado para o código: "
+                            + novaAfericaoPlaca.getCodDiagrama());
         }
         return ProtheusRodalogConverter.convertNovaAfericaoPlaca(novaAfericaoPlaca, diagramaVeiculo.get());
     }