+++ src/main/java/br/com/zalf/prolog/webservice/integracao/sistema/SistemasFactory.java
@@ -6,7 +6,7 @@ import br.com.zalf.prolog.webservice.integracao.avacorpavilan.requester.AvaCorpA
 import br.com.zalf.prolog.webservice.integracao.protheusrodalog.ProtheusRodalogRequesterImpl;
 import br.com.zalf.prolog.webservice.integracao.protheusrodalog.SistemaProtheusRodalog;
 import br.com.zalf.prolog.webservice.integracao.transport.SistemaTransportTranslecchi;
-import com.sun.istack.internal.NotNull;
+import org.jetbrains.annotations.NotNull;
 
 /**
  * Created by luiz on 7/17/17.
@@ -24,7 +24,11 @@ public final class SistemasFactory {
 
         switch (sistemaKey) {
             case AVACORP_AVILAN:
-                return new AvaCorpAvilan(new AvaCorpAvilanRequesterImpl(), sistemaKey, integradorProLog, userToken);
+                return new AvaCorpAvilan(
+                        new AvaCorpAvilanRequesterImpl(),
+                        sistemaKey,
+                        integradorProLog,
+                        userToken);
             case TRANSPORT_TRANSLECCHI:
                 return new SistemaTransportTranslecchi(
                         sistemaKey,