+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Change Log
 <a name="v0.0.93"></a>
 ## Version [v0.0.93](https://github.com/luizfp/PrologWebService/compare/v0.0.92...v0.0.93) (release-date) [unreleased]
 
+### Features
+* Cria integração de aferição de placas com o Protheus da Rodalog (PL-1721)
 
 <a name="v0.0.92"></a>
 ## Version [v0.0.92](https://github.com/luizfp/PrologWebService/compare/v0.0.91...v0.0.92) (2019-05-16)